refactor(tcflower): introduce named ipProto type
Make the IP protocol match a named ipProto type (with ipProtoAny/TCP/
UDP/SCTP) instead of a bare uint8, so the "0 = match any L4" contract is
explicit and proto values can't be mixed up with other small integers.
Convert to uint8 only at the go-tc KeyIPProto boundary.

The L2/L3 protocol numbers remain hardcoded with a UAPI citation rather
than sourced from golang.org/x/sys/unix: the pure translation layer must
compile on darwin for the golden tests, and ETH_P_IP is Linux-only.

+++ pkg/tcflower/engine_test.go
@@ -346,7 +346,7 @@ func TestSinglePortAndPortRange(t *testing.T) {
 	if so.Flower.KeyTCPDst == nil || *so.Flower.KeyTCPDst != 80 {
 		t.Errorf("single port: expected KeyTCPDst=80, got %+v", so.Flower.KeyTCPDst)
 	}
-	if so.Flower.KeyIPProto == nil || *so.Flower.KeyIPProto != ipProtoTCP {
+	if so.Flower.KeyIPProto == nil || *so.Flower.KeyIPProto != uint8(ipProtoTCP) {
 		t.Errorf("single port: expected ip_proto TCP")
 	}
 	ro := rangeR.toObject(1)
@@ -464,7 +464,7 @@ func TestToObjectSkipSwAndKeys(t *testing.T) {
 	if obj.Flower.Flags == nil || *obj.Flower.Flags != tc.SkipSw {
 		t.Errorf("expected skip_sw flag set")
 	}
-	if obj.Flower.KeyIPProto == nil || *obj.Flower.KeyIPProto != ipProtoTCP {
+	if obj.Flower.KeyIPProto == nil || *obj.Flower.KeyIPProto != uint8(ipProtoTCP) {
 		t.Errorf("expected ip_proto TCP")
 	}
 	if obj.Flower.KeyTCPDst == nil || *obj.Flower.KeyTCPDst != 443 {
